fix: add float64 around the bitDepth to integer range calculator

+++ main.go
@@ -78,7 +78,7 @@ func read(ws *websocket.Conn) string {
 
 func bitDepthToIntegerRange(bitDepth int) float64 {
 	// 2^(bitDepth - 1) (the -1 is because it's signed lol)
-	return 1 << (bitDepth - 1)
+	return float64(1 << (bitDepth - 1))
 }
 
 func frequencyToPeriod(frequency float64) time.Duration {
